+++ tests/test_defaults.py
@@ -245,3 +245,48 @@ def test_wait_for_calc_applies_dynamically(backend, mongetter):
     res4 = _wait_for_calc_timeout_slow(1, 2)
     # One of the cached values is returned
     assert res1 == res4 or res2 == res4 or res3 == res4
+
+
+def test_default_kwargs_handling():
+    count = 0
+
+    @cachier.cachier()
+    def dummy_func(a, b=2):
+        nonlocal count
+        count += 1
+        return a + b
+
+    dummy_func.clear_cache()
+    assert count == 0
+    assert dummy_func(1) == 3
+    assert dummy_func(a=1) == 3
+    assert dummy_func(a=1, b=2) == 3
+    assert count == 1
+
+
+def test_deprecated_func_kwargs():
+    count = 0
+
+    @cachier.cachier()
+    def dummy_func(a, b=2):
+        nonlocal count
+        count += 1
+        return a + b
+
+    dummy_func.clear_cache()
+    assert count == 0
+    with pytest.deprecated_call(
+        match="`verbose_cache` is deprecated and will be removed"
+    ):
+        assert dummy_func(1, verbose_cache=True) == 3
+    assert count == 1
+    with pytest.deprecated_call(
+        match="`ignore_cache` is deprecated and will be removed"
+    ):
+        assert dummy_func(1, ignore_cache=True) == 3
+    assert count == 2
+    with pytest.deprecated_call(
+        match="`overwrite_cache` is deprecated and will be removed"
+    ):
+        assert dummy_func(1, overwrite_cache=True) == 3
+    assert count == 3
